identityserver

+++ src/AuctionService/Controllers/AuctionsController.cs
@@ -4,6 +4,7 @@ using AutoMapper.QueryableExtensions;
 using Contracts;
 using MassTransit;
 using MassTransit.DependencyInjection.Registration;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
@@ -47,11 +48,13 @@ public class AuctionsController:ControllerBase
 
         return _mapper.Map<AuctionDto>(auction);
     }
+
+    [Authorize]
     [HttpPost]
     public async Task<ActionResult<AuctionDto>> CreateAuction(CreateAuctionDto auctionDto)
     {
         var auction = _mapper.Map<Auction>(auctionDto);
-        auction.Seller="test";
+        auction.Seller=User.Identity.Name;
 
         _context.Auctions.Add(auction);
         var newAuction = _mapper.Map<AuctionDto>(auction);
@@ -63,7 +66,7 @@ public class AuctionsController:ControllerBase
         return CreatedAtAction(nameof(GetAuctionById),
             new { auction.Id }, newAuction);
     }
-
+    [Authorize]
      [HttpPut("{id}")]
     public async Task<ActionResult> UpdateAuction(Guid id, UpdateAuctionDto updateAuctionDto)
     {
@@ -71,8 +74,8 @@ public class AuctionsController:ControllerBase
 
         if (auction == null) return NotFound();
         
-        //ToDO: check seller for updating
-        // if (auction.Seller != User.Identity.Name) return Forbid();
+        if (auction.Seller != User.Identity.Name) return Forbid();
+
         auction.Item.Make = updateAuctionDto.Make ?? auction.Item.Make;
         auction.Item.Model = updateAuctionDto.Model ?? auction.Item.Model;
         auction.Item.Color = updateAuctionDto.Color ?? auction.Item.Color;
@@ -87,7 +90,7 @@ public class AuctionsController:ControllerBase
 
         return BadRequest("Problem saving changes");
     }
-
+    [Authorize]
      [HttpDelete("{id}")]
     public async Task<ActionResult> DeleteAuction(Guid id)
     {
@@ -95,8 +98,7 @@ public class AuctionsController:ControllerBase
 
         if (auction == null) return NotFound();
         
-        //TODO: Check seller == username
-        // if (auction.Seller != User.Identity.Name) return Forbid();
+        if (auction.Seller != User.Identity.Name) return Forbid();
 
         _context.Auctions.Remove(auction);
